+++ src/main/java/shop/mtcoding/projectjobplan/apply/ApplyController.java
@@ -1,14 +1,11 @@
 package shop.mtcoding.projectjobplan.apply;
 
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import shop.mtcoding.projectjobplan._core.utils.ApiUtil;
 import shop.mtcoding.projectjobplan.user.SessionUser;
-import shop.mtcoding.projectjobplan.user.User;
 
 @RequiredArgsConstructor
 @RestController
@@ -31,9 +28,8 @@ public class ApplyController {
         return ResponseEntity.ok(new ApiUtil(applyDTO));
     }
 
-    @PutMapping("/api/applies") // /api/users/{userId}/applies
-    public ResponseEntity<?> update(@RequestBody ApplyRequest.UpdateDTO requestDTO) { // 지원자 합격/불합격 처리
-//        SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
+    @PutMapping("/api/users/{userId}/applies") // 지원자 합격/불합격 처리
+    public ResponseEntity<?> update(@PathVariable int userId, @RequestBody ApplyRequest.UpdateDTO requestDTO) {
         session.getAttribute("sessionUser");
         ApplyResponse.UpdateDTO updateDTO = applyService.updateApply(requestDTO);
