+++ src/main/java/shop/mtcoding/projectjobplan/apply/ApplyJpaRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ApplyJpaRepository extends JpaRepository<Apply, Integer> {
