+++ src/main/java/shop/mtcoding/projectjobplan/board/BoardRequest.java
@@ -3,7 +3,6 @@ package shop.mtcoding.projectjobplan.board;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
-import shop.mtcoding.projectjobplan.user.SessionUser;
 import shop.mtcoding.projectjobplan.user.User;
 
 import java.util.ArrayList;
@@ -13,7 +12,7 @@ public class BoardRequest {
 
     @Data
     public static class SaveDTO {
-        @Size(min = 1,max = 20, message = "제목은 20자를 초과할 수 없습니다.")
+        @Size(min = 1, max = 20, message = "제목은 20자를 초과할 수 없습니다.")
         @NotEmpty
         private String title;
         @NotEmpty
@@ -32,7 +31,6 @@ public class BoardRequest {
         private List<String> skill = new ArrayList<>();
 
 
-
         public Board toEntity(User user) {
             return Board.builder()
                     .user(user)
@@ -49,7 +47,7 @@ public class BoardRequest {
 
     @Data
     public static class UpdateDTO {
-        @Size(min = 1,max = 20, message = "제목은 20자를 초과할 수 없습니다.")
+        @Size(min = 1, max = 20, message = "제목은 20자를 초과할 수 없습니다.")
         @NotEmpty
         private String title;
         @NotEmpty
