+++ src/main/java/shop/mtcoding/projectjobplan/offer/OfferController.java
@@ -1,14 +1,11 @@
 package shop.mtcoding.projectjobplan.offer;
 
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import shop.mtcoding.projectjobplan._core.utils.ApiUtil;
 import shop.mtcoding.projectjobplan.user.SessionUser;
-import shop.mtcoding.projectjobplan.user.User;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,20 +23,20 @@ public class OfferController {
     @PostMapping("/api/resumes/{resumeId}/offers")
     public ResponseEntity<?> offer(@PathVariable Integer resumeId, @RequestBody OfferRequest.OfferDTO requestDTO) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
-        OfferResponse.OfferDTO offerDTO= offerService.createOffer(requestDTO);
+        OfferResponse.OfferDTO offerDTO = offerService.createOffer(requestDTO);
         return ResponseEntity.ok(new ApiUtil(offerDTO));
     }
 
-    @PutMapping("/api/offers")
-    public ResponseEntity<?> update(@RequestBody OfferRequest.UpdateDTO requestDTO) {
+    @PutMapping("/api/users/{userId}/offers")
+    public ResponseEntity<?> update(@PathVariable int userId, @RequestBody OfferRequest.UpdateDTO requestDTO) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
         OfferResponse.UpdateDTO updateDTO = offerService.updateOffer(requestDTO);
 
         return ResponseEntity.ok(new ApiUtil(updateDTO));
     }
 
-    @DeleteMapping("/api/offers/{offerId}")
-    public ResponseEntity<?> delete(@PathVariable int offerId) {
+    @DeleteMapping("/api/users/{userId}/offers/{offerId}")
+    public ResponseEntity<?> delete(@PathVariable int userId, @PathVariable int offerId) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
         offerService.removeOffer(offerId);
         return ResponseEntity.ok(new ApiUtil(null));
