+++ src/main/java/shop/mtcoding/projectjobplan/rating/RatingController.java
@@ -3,14 +3,12 @@ package shop.mtcoding.projectjobplan.rating;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import shop.mtcoding.projectjobplan._core.utils.ApiUtil;
 import shop.mtcoding.projectjobplan.user.SessionUser;
-import shop.mtcoding.projectjobplan.user.User;
 
 @RequiredArgsConstructor
 @RestController
@@ -21,7 +19,7 @@ public class RatingController {
     @PostMapping("/api/boards/{boardId}/rating")
     public ResponseEntity<?> rateBoard(@PathVariable int boardId, @RequestBody RatingRequest.RateBoardUser requestDTO) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
-        RatingResponse.DTO respDTO  = ratingService.createRating(sessionUser, requestDTO);
+        RatingResponse.DTO respDTO = ratingService.createRating(sessionUser, requestDTO);
 
         return ResponseEntity.ok(new ApiUtil(respDTO));
     }
