+++ src/main/java/shop/mtcoding/projectjobplan/resume/ResumeRequest.java
@@ -8,7 +8,7 @@ public class ResumeRequest {
 
     @Data // 이력서 작성하기
     public static class PostDTO {
-        @Size(min = 1,max = 20, message = "제목은 20자를 초과할 수 없습니다.")
+        @Size(min = 1, max = 20, message = "제목은 20자를 초과할 수 없습니다.")
         @NotEmpty
         private String title;
         @NotEmpty
@@ -19,7 +19,7 @@ public class ResumeRequest {
 
     @Data
     public static class UpdateDTO {
-        @Size(min = 1,max = 20, message = "제목은 20자를 초과할 수 없습니다.")
+        @Size(min = 1, max = 20, message = "제목은 20자를 초과할 수 없습니다.")
         @NotEmpty
         private String title;
         @NotEmpty
