+++ src/main/java/shop/mtcoding/projectjobplan/skill/Skill.java
@@ -7,8 +7,6 @@ import lombok.NoArgsConstructor;
 import shop.mtcoding.projectjobplan.board.Board;
 import shop.mtcoding.projectjobplan.user.User;
 
-import java.util.List;
-
 @NoArgsConstructor
 @Data
 @Entity
