--- src/main/java/shop/mtcoding/projectjobplan/skill/SkillRequest.java
@@ -1,14 +1,0 @@
-package shop.mtcoding.projectjobplan.skill;
-
-import lombok.Data;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class SkillRequest {
-
-    @Data
-    public static class DTO {
-        private List<String> skill = new ArrayList<>();
-    }
-}
